# perf: optimize trace key generation with hash-based deduplication